fix series test

+++ bodo/tests/test_series_old.py
@@ -475,8 +475,8 @@ class TestSeries(unittest.TestCase):
             return S + B
 
         n = 11
-        A = pd.Series(np.arange(n))
-        B = pd.Series(np.arange(n) ** 2)
+        A = pd.Series(np.arange(n, dtype=np.int64))
+        B = pd.Series(np.arange(n, dtype=np.int64) ** 2)
         bodo_func = bodo.jit(test_impl)
         pd.testing.assert_series_equal(bodo_func(A, B), test_impl(A, B))
         self.assertEqual(count_parfor_REPs(), 3)
